Add files via upload
Virtual Pizza Boy is an Android application through which a customer can order pizzas and other food items and an admin can monitor the menu and customer orders. The app also allows the user to pay the bill through app itself. The system is flexible to use and reduces the work of user to go to the counter or wait in the waiting line and it also decreases the efforts of the staff since the menu is provided in the app and admin can manage it. The admin is also provided with the current order list.The application is user friendly and requires minimum user intervention